Seperate editor logic by type
