+++ resources/js/app.js
@@ -1,7 +1,7 @@
 import './components/wysiwyg/tiptap.js';
 import './components/card/card.js';
 import './components/sidebar/sidebar.js';
-import './components/editor/editor.js';
+import './components/editor/edit.js';
 import './misc/misc.js';
 import './components/search/search.js';
 import './init.js';
